+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/tasks/bundling/BootJar.java
@@ -22,16 +22,15 @@ import java.util.concurrent.Callable;
 
 import org.gradle.api.Action;
 import org.gradle.api.artifacts.Configuration;
+import org.gradle.api.artifacts.ResolvableDependencies;
 import org.gradle.api.file.CopySpec;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.FileCopyDetails;
 import org.gradle.api.file.FileTreeElement;
 import org.gradle.api.internal.file.copy.CopyAction;
-import org.gradle.api.plugins.JavaPluginConvention;
 import org.gradle.api.specs.Spec;
 import org.gradle.api.tasks.Internal;
 import org.gradle.api.tasks.Nested;
-import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.bundling.Jar;
 
 /**
@@ -55,6 +54,8 @@ public class BootJar extends Jar implements BootArchive {
 
 	private static final String CLASSPATH_INDEX = "BOOT-INF/classpath.idx";
 
+	private final ResolvedDependencies resolvedDependencies = new ResolvedDependencies();
+
 	private final BootArchiveSupport support;
 
 	private final CopySpec bootInfSpec;
@@ -73,6 +74,11 @@ public class BootJar extends Jar implements BootArchive {
 		this.bootInfSpec = getProject().copySpec().into("BOOT-INF");
 		configureBootInfSpec(this.bootInfSpec);
 		getMainSpec().with(this.bootInfSpec);
+		getProject().getConfigurations().all((configuration) -> {
+			ResolvableDependencies incoming = configuration.getIncoming();
+			incoming.afterResolve(
+					(resolvableDependencies) -> this.resolvedDependencies.processConfiguration(configuration));
+		});
 	}
 
 	private void configureBootInfSpec(CopySpec bootInfSpec) {
@@ -108,12 +114,7 @@ public class BootJar extends Jar implements BootArchive {
 	@Override
 	protected CopyAction createCopyAction() {
 		if (!isLayeredDisabled()) {
-			JavaPluginConvention javaPluginConvention = getProject().getConvention()
-					.findPlugin(JavaPluginConvention.class);
-			Iterable<SourceSet> sourceSets = (javaPluginConvention != null) ? javaPluginConvention.getSourceSets()
-					: Collections.emptySet();
-			LayerResolver layerResolver = new LayerResolver(sourceSets, getConfigurations(), this.layered,
-					this::isLibrary);
+			LayerResolver layerResolver = new LayerResolver(this.resolvedDependencies, this.layered, this::isLibrary);
 			String layerToolsLocation = this.layered.isIncludeLayerTools() ? LIB_DIRECTORY : null;
 			return this.support.createCopyAction(this, layerResolver, layerToolsLocation);
 		}
@@ -306,6 +307,11 @@ public class BootJar extends Jar implements BootArchive {
 	 */
 	private static <T> Callable<T> callTo(Callable<T> callable) {
 		return callable;
+	}
+
+	@Internal
+	ResolvedDependencies getResolvedDependencies() {
+		return this.resolvedDependencies;
 	}
 
 }
